cast mechanic refactoring
