+++ GW2EIEvtcParser/EIData/Mechanics/CastMechanics/CastMechanic.cs
@@ -1,4 +1,5 @@
-﻿using GW2EIEvtcParser.ParsedData;
+﻿using System.Collections.Generic;
+using GW2EIEvtcParser.ParsedData;
 
 namespace GW2EIEvtcParser.EIData
 {
@@ -16,6 +17,8 @@ namespace GW2EIEvtcParser.EIData
 
         protected abstract long GetTime(AbstractCastEvent evt);
 
+        protected abstract AbstractSingleActor GetActor(ParsedEvtcLog log, AgentItem agentItem, Dictionary<int, AbstractSingleActor> regroupedMobs);
+
         public CastMechanic(long mechanicID, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, CastChecker condition = null) : this(new long[] { mechanicID }, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
         {
         }
@@ -24,5 +27,25 @@ namespace GW2EIEvtcParser.EIData
         {
             _triggerCondition = condition;
         }
+
+        internal override void CheckMechanic(ParsedEvtcLog log, Dictionary<Mechanic, List<MechanicEvent>> mechanicLogs, Dictionary<int, AbstractSingleActor> regroupedMobs)
+        {
+            foreach (long mechanicID in MechanicIDs)
+            {
+                foreach (AbstractCastEvent c in log.CombatData.GetAnimatedCastData(mechanicID))
+                {
+                    if (Keep(c, log))
+                    {
+                        AbstractSingleActor amp = GetActor(log, c.Caster, regroupedMobs);
+                        if (amp != null)
+                        {
+                            mechanicLogs[this].Add(new MechanicEvent(GetTime(c), this, amp));
+                        }
+                    }
+                }
+            }
+
+        }
+
     }
 }
