+++ GW2EIEvtcParser/EIData/Mechanics/CastMechanics/Enemy/EnemyCastMechanic.cs
@@ -17,24 +17,9 @@ namespace GW2EIEvtcParser.EIData
             IsEnemyMechanic = true;
         }
 
-        internal override void CheckMechanic(ParsedEvtcLog log, Dictionary<Mechanic, List<MechanicEvent>> mechanicLogs, Dictionary<int, AbstractSingleActor> regroupedMobs)
+        protected override AbstractSingleActor GetActor(ParsedEvtcLog log, AgentItem agentItem, Dictionary<int, AbstractSingleActor> regroupedMobs)
         {
-            foreach (long mechanicID in MechanicIDs)
-            {
-                foreach (AbstractCastEvent c in log.CombatData.GetAnimatedCastData(mechanicID))
-                {
-                    AbstractSingleActor amp = null;
-                    if (Keep(c, log))
-                    {
-                        amp = MechanicHelper.FindEnemyActor(log, c.Caster, regroupedMobs);
-                    }
-                    if (amp != null)
-                    {
-                        mechanicLogs[this].Add(new MechanicEvent(GetTime(c), this, amp));
-                    }
-                }
-            }
-
+            return MechanicHelper.FindEnemyActor(log, agentItem, regroupedMobs);
         }
     }
 }
